// src/logging/log-type.enum.ts
/** Represents logger message's type. */
export enum LogType
{
    /** Specifies that a message is for debug purposes. */
    Debug = 0,

    /** Specifies that a message represents a simple information. */
    Information = 1,

    /** Specifies that a message represents a warning. */
    Warning = 2,

    /** Specifies that a message represents an error. */
    Error = 3
}